requirements.txt
flask
python-dotenv
openai
aiohttp
quart
requests
flask[async]
aiofiles